requirements.txt
absl-py==0.15.0
altair==4.1.0
appnope==0.1.2
argon2-cffi==21.1.0
astor==0.8.1
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
blis==0.4.1
boto3==1.19.7
botocore==1.22.7
bs4==0.0.1
cachetools==4.2.4
catalogue==1.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
ChatterBot==1.0.8
chatterbot-corpus==1.2.0
clang==5.0
click==8.0.1
cymem==2.0.5
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
en-core-web-sm==2.3.1
entrypoints==0.3
enum-compat==0.0.3
Flask==2.0.2
flatbuffers==1.12
gast==0.4.0
google-auth==2.3.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
h5py==3.1.0
idna==3.3
ipykernel==6.4.2
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.0.1
jsonschema==4.1.2
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==2.0.1
mathparse==0.1.2
matplotlib-inline==0.1.3
mistune==0.8.4
murmurhash==1.0.5
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
nltk==3.5
notebook==6.4.5
numpy==1.18.4
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pandas==1.0.4
pandocfilters==1.5.0
parso==0.8.2
pathlib==1.0.1
pathy==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
plac==1.1.3
preshed==3.0.5
prometheus-client==0.12.0
prompt-toolkit==3.0.21
protobuf==3.19.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
pydeck==0.7.1
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==3.13
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
s3transfer==0.5.0
Send2Trash==1.8.0
six==1.15.0
smart-open==5.2.1
soupsieve==2.3.1
spacy==2.3.1
spacy-legacy==3.0.8
SQLAlchemy==1.3.24
srsly==1.0.5
streamlit==0.62.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.6.0
tensorflow-estimator==2.7.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
thinc==7.4.1
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.62.2
traitlets==5.1.1
typer==0.4.0
typing-extensions==3.7.4.3
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
validators==0.18.2
wasabi==0.8.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.12.1
